src/requirements.txt
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.10.0
asttokens==3.0.1
async-generator==1.10
async-lru==2.3.0
attrs==24.2.0
Automat==24.8.1
babel==2.18.0
beautifulsoup4==4.11.2
bleach==6.3.0
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.0.1
chromedriver-py==145.0.7632.46
click==8.1.3
colorama==0.4.6
comm==0.2.3
comtypes==1.3.1
constantly==23.10.4
contourpy==1.3.3
cryptography==43.0.3
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.7
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.16.1
Flask==2.2.2
fonttools==4.62.1
fqdn==1.5.1
fsspec==2024.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.4
importlib_metadata==8.7.1
incremental==24.7.2
ipykernel==7.2.0
ipython==9.10.1
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.5.3
json5==0.14.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kagglehub==1.0.0
kagglesdk==0.1.15
kiwisolver==1.5.0
labeling==0.1.14
lark==1.3.1
libretranslatepy==2.1.1
lxml==5.3.0
MarkupSafe==2.1.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.2.0
MouseInfo==0.1.3
mpmath==1.3.0
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nmap==0.0.1
notebook==7.5.5
notebook_shim==0.2.4
numpy==1.26.4
opencv-python==4.11.0.86
outcome==1.3.0.post0
overrides==7.7.0
packaging==26.0
pandas==3.0.1
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.4
prometheus_client==0.24.1
prompt_toolkit==3.0.52
Protego==0.3.1
protobuf==6.33.5
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.53
pycparser==2.22
PyDispatcher==2.0.7
PyGetWindow==0.0.9
Pygments==2.20.0
pykwalify==1.8.0
PyMsgBox==1.0.9
pyOpenSSL==24.2.1
pyparsing==3.3.2
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
pyradiomics==3.0.1
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.7
python-dotenv==0.21.1
python-json-logger==4.1.0
python-nmap==0.7.1
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.5
PyWavelets==1.9.0
pywhatkit==5.4
pywin32==306
pywinpty==3.0.3
PyYAML==6.0.2
pyzmq==27.1.0
queuelib==1.7.0
referencing==0.37.0
requests==2.32.5
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
ruamel.yaml==0.19.1
scikit-learn==1.8.0
scipy==1.17.1
Scrapy==2.12.0
seaborn==0.13.2
selenium==4.40.0
Send2Trash==2.1.0
service-identity==24.2.0
simpleitk==2.5.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.10.1
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
TA-Lib @ file:///C:/Users/nanda/Downloads/TA_Lib-0.4.28-cp311-cp311-win_amd64.whl/TA_Lib-0.4.28-cp311-cp311-win_amd64.whl#sha256=a68761dcbea7e1346c9c0e544da74f0bb12b324f8ae363bb27f126f36fe47254
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.1.10
tinycss2==1.4.0
tk==0.1.0
tldextract==5.1.3
torch==2.5.1
torchvision==0.20.1
tornado==6.5.5
tqdm==4.64.1
traitlets==5.14.3
translate==3.6.1
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
Twisted==24.10.0
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.66
ultralytics-thop==2.0.14
undetected-chromedriver==3.5.5
uri-template==1.3.0
urllib3==2.6.3
w3lib==2.2.1
wcwidth==0.6.0
webcolors==25.10.0
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==2.2.2
wikipedia==1.4.0
wsproto==1.3.2
zipp==3.23.0
zope.interface==7.1.1
